backend: Add translated / untranslated tab in menu
Checkbox was replaced by tab control with 'translated' and 'to translation' elements.
Added new records for menu table (to see if divide by translated / untranslated records works on backent).
Replaced identifier in table for records (in backend) to index of element (originally was order, which may be differend in other languages).