core: Add database migrations
Added sequelize as SQL Query builder for MySQL/MariaDB, PostgreSQL and SQLite.
SQLite is not recommended for this project, hovewer it will work.
Migrations are partly version control for database.
Need to check if CHAR(16) in BINARY form is ok for ID fields (originaly should be BINARY(16)).
Replaced ini files for languages to php files.