fm: Create UI for FileManager
Starting UI for FileManager, with buttons for refresh, create file, folder and go to root folder.
In top bar on second panel there will be current location, for root it's name will be highlighted on red.
On left panel there will be folder tree, on right files and folders available for current location.

Issue: #43