core: Display menus in left sidebar
This commit render menu names in left sidebar with link to edit.
To do this, all controllers that are displaying data should extends from BaseController located in Helpers directory.
Fixed prepare.js - it was trying to minify RequireJS, but it was removed by previous commit.
Divided Volt code to partials (pages and breadcrumbs).

Issue: #46