core: Display folders in file manager
Added two dependencies for static version - doT (~3kb) and qwest (~8kb) for templates and ajax.
FileManager must be dynamic even in static version of website, so, for simplify process of displaying folders and entities created Observable and Render objects.
Folders are displaying in tree, for this purpose created RenderArrayTree object.
FilemanagerController for API now uses FilemanagerService (which will be shared between controllers).

Issues: #43, #47