junk: Clean code add logging
This commit adds Logger class for JavaScript.
It helps to logging errors to console and variables (can be easly turned off).
Cleaning code - all TypeScript files was cleaned, code was corrected.
Added comments and comment header for files in which it was missing.
Rewrite concept for catching elements in filemanager - now it operaters on data attributes, not on ID.
Cleaned filemanager index view.
Corrected css classes.

Issue: #43